@@ -26,14 +26,14 @@ final class HeaderItemHelper extends Attributes implements ItemHelper
      *
      * @var array
      */
-    protected $item;
+    protected array $item;
 
     /**
      * Item classes.
      *
      * @var array
      */
-    protected $itemClass = array();
+    protected array $itemClass = array();
 
     /**
      * AbstractItemHelper constructor.
